refactor: separate the initializing point for app
