+++ db.js
@@ -1,0 +1,10 @@
+const { sequelize } = require("./models");
+
+sequelize
+  .sync({ force: false })
+  .then(() => {
+    console.log("DB 연결 되었습니다.");
+  })
+  .catch((err) => {
+    console.log(err);
+  });
